requirements.txt
# ============================================================
# Agentic AI IT Support Automation System
# Requirements
# ============================================================

# ---- Web Backend ----
fastapi==0.115.5
uvicorn[standard]==0.32.1
jinja2==3.1.4
python-multipart==0.0.18  # Required for FastAPI form parsing

# ---- AI / LLM ----
groq==0.12.0               # Groq Python SDK for fast LLM inference

# ---- Browser Automation ----
playwright==1.49.0          # Playwright for Python (browser control)

# ---- Environment & Config ----
python-dotenv==1.0.1        # Load .env files

# ---- Utilities ----
# (No heavy optional deps — keeping it lean)
